// src/common/enums/role.enum.ts
export enum Role {
  SUPER_ADMIN = "super_admin",
  SYSTEM_ADMIN = "system_admin",
  DATA_ENTRY_OFFICER = "data_entry_officer",
  DATA_MANAGER = "data_manager",
  VILLAGE_OFFICER = "village_officer",
  WELFARE_OFFICER = "welfare_officer",
  AGRICULTURE_OFFICER = "agriculture_officer",
  HEALTH_OFFICER = "health_officer",
  EDUCATION_OFFICER = "education_officer",
  ANALYST = "analyst",
  AUDITOR = "auditor",
  CITIZEN = "citizen"
}